requirements.txt
flask==3.1.0
supabase==2.15.0
anthropic==0.49.0
python-dotenv==1.0.1
